requirements.txt
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.34.0
anyio==4.13.0
asyncpg==0.29.0
certifi==2026.2.25
click==8.3.1
distro==1.9.0
fastapi==0.115.0
filelock==3.25.2
fsspec==2026.3.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
huggingface_hub==1.8.0
idna==3.11
jiter==0.13.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.20.0
python-dotenv==1.0.1
PyYAML==6.0.3
rich==14.3.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
